[#20] refactor: dummyDebateList dummyData로 분리
