+++ src/app/debateList/page.tsx
@@ -1,29 +1,9 @@
 import ListCard from '@/components/Common/ListCard';
 import Wrapper from '@/components/Common/Wrapper';
+import dummyDebateList from '@/constants/dummyData';
 import React from 'react';
 
 const page = () => {
-    const dummyDebateList = [
-        {
-            id: 1,
-            title: '마리모는 식물이 아니라 동물입니다',
-            username: '독수리타법 7남매',
-            time: '2024. 01. 10. 11:38',
-            content:
-                '마리모는 동물입니다. 동물을 식물이라고 부르는 것은 마리모에게 실례입니다.',
-            likeNum: 3,
-            dislikeNum: 1,
-        },
-        {
-            id: 2,
-            title: '오늘의 날씨는 궁금하지가 않습니다',
-            username: '여행갈고양',
-            time: '2024. 01. 10. 11:22',
-            content: '여행갈 때에는 오늘의 날씨가 중요하긴 합니다. 그러나',
-            likeNum: 22,
-            dislikeNum: 3,
-        },
-    ];
     return (
         <Wrapper>
             {dummyDebateList.map(item => {
